% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cat_results}
\alias{cat_results}
\title{print scan results}
\usage{
cat_results(res)
}
\arguments{
\item{x}{output of \code{snif_secrets}}
}
\value{
cat'ed output to console
}
\description{
print scan results
}
\keyword{internal}
