% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snif_one.R
\name{snif_one}
\alias{snif_one}
\title{scan one secret string across many paths}
\usage{
snif_one(path, secret_string)
}
\arguments{
\item{path}{(character) path to fixtures directory. required}

\item{secret_string}{(character) a character string to look for. required}
}
\value{
a named list, with vector of line numbers for where string found;
if none found an empty list
}
\description{
scan one secret string across many paths
}
\examples{
\dontrun{
Sys.setenv(A_KEY = "a8d#d\%d7g7g4012a4s2")
path <- file.path(tempdir(), "foobar")
dir.create(path)
# no matches
snif_one(path, Sys.getenv("A_KEY"))
# add files with the secret
cat(paste0("foo\nbar\nhello\nworld\n", 
  Sys.getenv("A_KEY"), "\n"), file = file.path(path, "stuff.R"))
# matches!
snif_one(path, Sys.getenv("A_KEY"))
}
}
