% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snif_secrets_pkg.R
\name{snif_secrets_pkg}
\alias{snif_secrets_pkg}
\title{Scan for secrets in a package}
\usage{
snif_secrets_pkg(dir = ".", secrets)
}
\arguments{
\item{dir}{(character) path to a package root. required}

\item{secrets}{(character) vector of secrets to look for. required}
}
\value{
named list of all secrets, each either an empty list if none
found, or a named list of files and the line numbers secrets found on
}
\description{
Scan for secrets in a package
}
\examples{
\dontrun{
pkgpath = tempdir()
snif_secrets_pkg(dir = pkgpath, secrets = c("mysecretkey"))
unlink(pkgpath)
}
}
