% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snif_secrets_fixtures.R
\name{sniff_secrets_fixtures}
\alias{sniff_secrets_fixtures}
\title{Scan for secrets in a package's fixtures}
\usage{
sniff_secrets_fixtures(dir = ".")
}
\arguments{
\item{dir}{path to a package root}
}
\value{
named list of all secrets, each either an empty list if none
found, or a named list of files and the line numbers secrets found on
}
\description{
Scan for secrets in a package's fixtures
}
\examples{
\dontrun{
# dir = tempdir()
path <- "/Users/sckott/github/ropensci/taxize"
sniff_secrets_fixtures(dir = path)
}
}
