% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniffer.R
\name{sniffer}
\alias{sniffer}
\title{sniffer - pretty print method}
\usage{
sniffer(dir = ".")
}
\arguments{
\item{dir}{path to a package root}
}
\value{
named list of all secrets, each either an empty list if none
found, or a named list of files and the line numbers secrets found on
}
\description{
sniffer - pretty print method
}
\examples{
\dontrun{
# dir = tempdir()
path <- "/Users/sckott/github/ropensci/taxize"
sniffer(dir = path)
}
}
