% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trufflesniffer-package.R
\docType{package}
\name{trufflesniffer-package}
\alias{trufflesniffer-package}
\alias{trufflesniffer}
\title{sniff out secrets}
\description{
sniff out secrets
}
\author{
Scott Chamberlain \href{mailto:sckott@protonmail.com}{sckott@protonmail.com}
}
\keyword{package}
